`get_results()` returns Empty Array if no result
Previously, a call to get_results() would return NULL if the output was
ARRAY and no results were available. I have changed this to return an
empty array (as this is the expected response type, is falsely, and
prevents type errors when then output it used in a foreach loop.